Bugfix: Show Calendar if unable to determine Location
