+++ src/components/InfoWidget/InfoWidget.tsx
@@ -1,19 +1,19 @@
 import { useQuery } from "@tanstack/react-query";
 
 import { getLocation } from "@/src/services/location";
-import { getWeather, mapWeatherCode } from "@/src/services/weather";
-import { cn, formatToday, formatWeekNumber } from "@/src/utils";
+import { getWeather } from "@/src/services/weather";
+import { cn } from "@/src/utils";
 
-import ErrorState from "./partials/ErrorState";
+import CalendarWidget from "./partials/CalendarWidget";
 import LoadingState from "./partials/LoadingState";
-import Link from "../ui/Link";
+import WeatherWidget from "./partials/WeatherWidget";
 
 type Props = {
   className?: string;
 };
 
 const InfoWidget = () => {
-  const { data: locationData } = useQuery({
+  const { isError: isLocationError, data: locationData } = useQuery({
     queryFn: getLocation,
     queryKey: ["location"],
   });
@@ -32,46 +32,25 @@ const InfoWidget = () => {
     queryKey: ["weather", locationData?.latitude, locationData?.longitude],
   });
 
+  if (isLocationError) {
+    return <CalendarWidget />;
+  }
+
   if (isPending) {
     return <LoadingState />;
   }
 
   if (isError) {
-    return <ErrorState />;
+    return <span>Unable to display weather information.</span>;
   }
-
-  const weatherDescription = mapWeatherCode(
-    String(weatherData.current.weather_code)
-  );
-  const today = new Date();
 
   return (
     <>
-      <Link
-        href="https://www.google.com/search?q=weather"
-        className="flex items-center gap-6"
-      >
-        <p className="text-4xl font-bold">
-          {weatherData.current.temperature_2m}
-          {weatherData.current_units.temperature_2m}
-        </p>
-        {weatherDescription && (
-          <div className="flex items-center gap-2">
-            <div className="flex-col">
-              <p className="font-medium">{locationData?.area}</p>
-              <p className="text-xs opacity-70">
-                {weatherDescription.description}
-              </p>
-            </div>
-            <img className="w-10" src={weatherDescription.image} />
-          </div>
-        )}
-      </Link>
+      {weatherData && (
+        <WeatherWidget weatherData={weatherData} area={locationData?.area} />
+      )}
       <div className="h-10 w-0.5 rounded bg-black/5" />
-      <Link href="https://calendar.google.com" className="flex flex-col">
-        <p className="font-medium">{formatToday(today)}</p>
-        <p className="text-xs opacity-70">{formatWeekNumber(today)}</p>
-      </Link>
+      <CalendarWidget />
     </>
   );
 };
