--- src/components/InfoWidget/partials/ErrorState.tsx
@@ -1,3 +1,0 @@
-const ErrorState = () => <span>Unable to display weather information.</span>;
-
-export default ErrorState;
