+++ src/services/location/checkCachedLatLong.ts
@@ -4,9 +4,7 @@ import { Location } from "./types";
 
 const CACHE_DURATION = 20 * 60 * 1000; // 20 minutes
 
-/**
- * Persists the user's lat/long for 20 minutes to speed up performance
- */
+/* Persists the user's lat/long to speed up performance */
 export const checkCachedLatLong = async (): Promise<Location | void> => {
   const { cachedLocation } = await storage.local.get("cachedLocation");
